requirements.txt
asgiref==3.11.0
Django==5.2
django-cid==3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
ulid-py==1.1.0
